+++ src/server/OFIRMRES.cpp
@@ -113,7 +113,7 @@ void FirmRes::load_firm_bitmap()
 	FirmBitmapRec  *firmBitmapRec;
 	FirmBitmap     *firmBitmap;
 	int      		i;
-	long				bitmapOffset;
+	uint32_t		   bitmapOffset;
 	Database 		*dbFirmBitmap = game_set.open_db(FIRM_BITMAP_DB);
 
 	firm_bitmap_count = (short) dbFirmBitmap->rec_count();
@@ -128,7 +128,7 @@ void FirmRes::load_firm_bitmap()
 		firmBitmapRec = (FirmBitmapRec*) dbFirmBitmap->read(i+1);
 		firmBitmap    = firm_bitmap_array+i;
 
-		memcpy( &bitmapOffset, firmBitmapRec->bitmap_ptr, sizeof(long) );
+		memcpy( &bitmapOffset, firmBitmapRec->bitmap_ptr, sizeof(uint32_t) );
 
 		firmBitmap->bitmap_ptr = res_bitmap.read_imported(bitmapOffset);
 		firmBitmap->width  	  = *((short*)firmBitmap->bitmap_ptr);
